requirements.txt
python-dotenv==1.1.0
fastapi[standard]
openai
slowapi
redis